visa_docs_analysis/backend/pyproject.toml
[tool.poetry]
name = "visa_docs_analysis_backend"
version = "0.1.0"
description = "FastAPI backend for Visa Docs Analysis (migrated from Streamlit)"
authors = ["Alexey Nikitin <you@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.111.0"
uvicorn = {version = "^0.30.0", extras = ["standard"]}
pydantic = "^2.7.0"
pydantic-settings = "^2.3.0"
SQLAlchemy = "^2.0.30"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
httpx = "^0.27.0"
tenacity = "^8.2.3"
structlog = "^24.1.0"
python-dotenv = "^1.0.1"
aiosqlite = "^0.20.0"
celery = "^5.4.0"
redis = {version = "^5.0.7", extras = ["hiredis"]}
greenlet = "^3.0.3"
google-genai = "^0.6.0"
google-api-python-client = "^2.146.0"
google-auth = "^2.35.0"
google-auth-oauthlib = "^1.2.1"
python-docx = "^1.1.2"
PyPDF2 = "^3.0.1"
cryptography = "^43.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.6"
ruff = "^0.5.0"
mypy = "^1.10.0"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
  "E", "F", "W",        # pycodestyle/pyflakes
  "I",                    # isort
  "C90",                  # mccabe
  "UP",                   # pyupgrade
  "B",                    # bugbear
]
ignore = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-q"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

